refactor: other non-test to template
