+++ autonomy/cli/scaffolding_templates/__init__.py
@@ -20,6 +20,7 @@ FILE_HEADER = """\
 # ------------------------------------------------------------------------------
 """
 
+
 # Rounds
 ROUNDS_FILE_HEADER = """\
 \"\"\"This package contains the rounds of {AbciApp}.\"\"\"
@@ -94,7 +95,6 @@ class {RoundCls}({ABCRoundCls}):
 
 """
 
-
 ABCI_APP_CLS_TEMPLATE = """\
 class {AbciAppCls}(AbciApp[Event]):
     \"\"\"{AbciAppCls}\"\"\"
@@ -107,8 +107,9 @@ class {AbciAppCls}(AbciApp[Event]):
     cross_period_persisted_keys: List[str] = []
 """
 
+
 # Behaviours
-BEHAVIOUR_FILE_HEADER ="""\
+BEHAVIOUR_FILE_HEADER = """\
 \"\"\"This package contains round behaviours of {AbciApp}.\"\"\"
 
 from abc import abstractmethod
@@ -129,7 +130,6 @@ from {author}.skills.{skill_name}.rounds import (
 
 """
 
-
 BASE_BEHAVIOUR_CLS_TEMPLATE = """\
 class {BaseBehaviourCls}(BaseBehaviour):
     \"\"\"Base behaviour for the common apps' skill.\"\"\"
@@ -146,7 +146,6 @@ class {BaseBehaviourCls}(BaseBehaviour):
 
 """
 
-
 BEHAVIOUR_CLS_TEMPLATE = """\
 class {BehaviourCls}({BaseBehaviourCls}):
     \"\"\"{BehaviourCls}\"\"\"
@@ -162,7 +161,6 @@ class {BehaviourCls}({BaseBehaviourCls}):
 
 """
 
-
 ROUND_BEHAVIOUR_CLS_TEMPLATE = """\
 class {RoundBehaviourCls}(AbstractRoundBehaviour):
     \"\"\"{RoundBehaviourCls}\"\"\"
@@ -170,4 +168,185 @@ class {RoundBehaviourCls}(AbstractRoundBehaviour):
     initial_behaviour_cls = {InitialBehaviourCls}
     abci_app_cls = {AbciAppCls}  # type: ignore
     behaviours: Set[Type[BaseBehaviour]] = {behaviours}
+"""
+
+
+# Payloads
+PAYLOADS_FILE = """\
+    \"\"\"This module contains the transaction payloads of the {FSMName}.\"\"\"
+
+    from abc import ABC
+    from enum import Enum
+    from typing import Any, Dict, Hashable, Optional
+
+    from packages.valory.skills.abstract_round_abci.base import BaseTxPayload
+
+"""
+
+TRANSACTION_TYPE_SECTION = """\
+class TransactionType(Enum):
+    \"\"\"Enumeration of transaction types.\"\"\"
+
+    # TODO: define transaction types: e.g. TX_HASH: "tx_hash"
+    {tx_types}
+
+    def __str__(self) -> str:
+        \"\"\"Get the string value of the transaction type.\"\"\"
+        return self.value
+
+"""
+
+BASE_PAYLOAD_CLS = """\
+class Base{FSMName}Payload(BaseTxPayload, ABC):
+    \"\"\"Base payload for {FSMName}.\"\"\"
+
+    def __init__(self, sender: str, content: Hashable, **kwargs: Any) -> None:
+        \"\"\"Initialize a 'select_keeper' transaction payload.\"\"\"
+
+        super().__init__(sender, **kwargs)
+        setattr(self, f"_{{self.transaction_type}}", content)
+        p = property(lambda s: getattr(self, f"_{{self.transaction_type}}"))
+        setattr(self.__class__, f"{{self.transaction_type}}", p)
+
+    @property
+    def data(self) -> Dict[str, Hashable]:
+        \"\"\"Get the data.\"\"\"
+        return {{str(self.transaction_type): getattr(self, str(self.transaction_type))}}
+
+"""
+
+PAYLOAD_CLS_TEMPLATE = """\
+class {PayloadCls}(Base{FSMName}Payload):
+    \"\"\"Represent a transaction payload for the {RoundCls}.\"\"\"
+
+    # TODO: specify the transaction type
+    transaction_type = TransactionType.{tx_type}
+
+"""
+
+
+# Models
+MODEL_FILE_TEMPLATE = """\
+\"\"\"This module contains the shared state for the abci skill of {AbciApp}.\"\"\"
+
+from typing import Any
+
+from packages.valory.skills.abstract_round_abci.models import BaseParams
+from packages.valory.skills.abstract_round_abci.models import Requests as BaseRequests
+from packages.valory.skills.abstract_round_abci.models import (
+    SharedState as BaseSharedState,
+)
+from {author}.skills.{skill_name}.rounds import {AbciApp}
+
+
+class SharedState(BaseSharedState):
+    \"\"\"Keep the current shared state of the skill.\"\"\"
+
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        \"\"\"Initialize the state.\"\"\"
+        super().__init__(*args, abci_app_cls={AbciApp}, **kwargs)
+
+
+Params = BaseParams
+Requests = BaseRequests
+"""
+
+
+# Handlers
+HANDLERS_FILE = """"\
+\"\"\"This module contains the handlers for the skill of {FSMName}.\"\"\"
+
+from packages.valory.skills.abstract_round_abci.handlers import (
+    ABCIRoundHandler as BaseABCIRoundHandler,
+)
+from packages.valory.skills.abstract_round_abci.handlers import (
+    ContractApiHandler as BaseContractApiHandler,
+)
+from packages.valory.skills.abstract_round_abci.handlers import (
+    HttpHandler as BaseHttpHandler,
+)
+from packages.valory.skills.abstract_round_abci.handlers import (
+    LedgerApiHandler as BaseLedgerApiHandler,
+)
+from packages.valory.skills.abstract_round_abci.handlers import (
+    SigningHandler as BaseSigningHandler,
+)
+from packages.valory.skills.abstract_round_abci.handlers import (
+    TendermintHandler as BaseTendermintHandler,
+)
+
+
+ABCIRoundHandler = BaseABCIRoundHandler
+HttpHandler = BaseHttpHandler
+SigningHandler = BaseSigningHandler
+LedgerApiHandler = BaseLedgerApiHandler
+ContractApiHandler = BaseContractApiHandler
+TendermintHandler = BaseTendermintHandler
+"""
+
+
+# Dialogues
+DIALOGUES_FILE = """\
+\"\"\"This module contains the dialogues of the {FSMName}.\"\"\"
+
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    AbciDialogue as BaseAbciDialogue,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    AbciDialogues as BaseAbciDialogues,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    ContractApiDialogue as BaseContractApiDialogue,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    ContractApiDialogues as BaseContractApiDialogues,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    HttpDialogue as BaseHttpDialogue,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    HttpDialogues as BaseHttpDialogues,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    LedgerApiDialogue as BaseLedgerApiDialogue,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    LedgerApiDialogues as BaseLedgerApiDialogues,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    SigningDialogue as BaseSigningDialogue,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    SigningDialogues as BaseSigningDialogues,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    TendermintDialogue as BaseTendermintDialogue,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    TendermintDialogues as BaseTendermintDialogues,
+)
+
+
+AbciDialogue = BaseAbciDialogue
+AbciDialogues = BaseAbciDialogues
+
+
+HttpDialogue = BaseHttpDialogue
+HttpDialogues = BaseHttpDialogues
+
+
+SigningDialogue = BaseSigningDialogue
+SigningDialogues = BaseSigningDialogues
+
+
+LedgerApiDialogue = BaseLedgerApiDialogue
+LedgerApiDialogues = BaseLedgerApiDialogues
+
+
+ContractApiDialogue = BaseContractApiDialogue
+ContactApiDialogues = BaseContractApiDialogues
+
+
+TendermintDialogue = BaseTendermintDialogue
+TendermintDialogues = BaseTendermintDialogues
 """
